Add function name validation
