+++ lib/serve/index.ts
@@ -21,6 +21,7 @@ import { FastApiContainer } from '../api-base/fastApiContainer';
 import { LisaServeApplicationConstruct, LisaServeApplicationProps } from './serveApplicationConstruct';
 import { ITable } from 'aws-cdk-lib/aws-dynamodb';
 export * from './serveApplicationConstruct';
+export * from './mcpWorkbenchConstruct';
 
 /**
  * LisaServe Application stack.
