+++ lib/serve/mcpWorkbenchConstruct.ts
@@ -18,33 +18,36 @@ import { IAuthorizer, IRestApi, RestApi } from 'aws-cdk-lib/aws-apigateway';
 import { ISecurityGroup } from 'aws-cdk-lib/aws-ec2';
 import { Construct } from 'constructs';
 import { Vpc } from '../networking/vpc';
-import { BaseProps, Config } from '../schema';
+import { BaseProps, Config, EcsSourceType } from '../schema';
 import * as s3 from 'aws-cdk-lib/aws-s3';
-import { RemovalPolicy, StackProps } from 'aws-cdk-lib';
+import { Duration, RemovalPolicy, StackProps } from 'aws-cdk-lib';
 import { createCdkId } from '../core/utils';
 import * as ssm from 'aws-cdk-lib/aws-ssm';
 import { getDefaultRuntime, PythonLambdaFunction, registerAPIEndpoint } from '../api-base/utils';
 import * as iam from 'aws-cdk-lib/aws-iam';
-import { LAMBDA_PATH } from '../util';
+import { LAMBDA_PATH, MCP_WORKBENCH_PATH } from '../util';
 import * as lambda from 'aws-cdk-lib/aws-lambda';
+import * as events from 'aws-cdk-lib/aws-events';
+import * as targets from 'aws-cdk-lib/aws-events-targets';
+import { ECSCluster, ECSTasks } from '../api-base/ecsCluster';
+import { Ec2Service } from 'aws-cdk-lib/aws-ecs';
 
 export type McpWorkbenchConstructProps = {
-    authorizer: IAuthorizer;
     restApiId: string;
     rootResourceId: string;
     securityGroups: ISecurityGroup[];
     vpc: Vpc;
+    apiCluster: ECSCluster;
+    authorizer?: IAuthorizer;
 } & BaseProps & StackProps;
 
-export default class McpWorkbenchConstruct extends Construct {
+export class McpWorkbenchConstruct extends Construct {
     public readonly workbenchBucket: s3.Bucket;
 
     constructor (scope: Construct, id: string, props: McpWorkbenchConstructProps) {
         super(scope, id);
 
-        const { authorizer, config, restApiId, rootResourceId, securityGroups, vpc } = props;
-
-        const workbenchBucket = this.createWorkbenchBucket(scope, config);
+        const { authorizer, config, restApiId, rootResourceId, securityGroups, vpc, apiCluster } = props;
 
         // Get common layer based on arn from SSM due to issues with cross stack references
         const commonLambdaLayer = lambda.LayerVersion.fromLayerVersionArn(
@@ -64,14 +67,14 @@ export default class McpWorkbenchConstruct extends Construct {
             rootResourceId: rootResourceId,
         });
 
-        const lambdaLayers = [commonLambdaLayer,  fastapiLambdaLayer];
+        const lambdaLayers = [commonLambdaLayer, fastapiLambdaLayer];
 
-        this.createWorkbenchApi(restApi, rootResourceId, config, vpc, securityGroups, authorizer, workbenchBucket, lambdaLayers);
-
-        this.workbenchBucket = workbenchBucket;
+        const workbenchBucket = this.createWorkbenchBucket(scope, config);
+        this.createWorkbenchApi(restApi, config, vpc, securityGroups, workbenchBucket, lambdaLayers, authorizer);
+        this.createWorkbenchService(apiCluster, config);
     }
 
-    private createWorkbenchApi (restApi: IRestApi, rootResourceId: string, config: Config, vpc: Vpc, securityGroups: ISecurityGroup[], authorizer: IAuthorizer, workbenchBucket: s3.Bucket, lambdaLayers: lambda.ILayerVersion[]) {
+    private createWorkbenchApi (restApi: IRestApi, config: Config, vpc: Vpc, securityGroups: ISecurityGroup[], workbenchBucket: s3.Bucket, lambdaLayers: lambda.ILayerVersion[], authorizer?: IAuthorizer) {
 
         const env = {
             ADMIN_GROUP: config.authConfig?.adminGroup || '',
@@ -187,5 +190,123 @@ export default class McpWorkbenchConstruct extends Construct {
             serverAccessLogsPrefix: 'logs/mcpworkbench-bucket/',
             eventBridgeEnabled: true
         });
+    }
+
+    private createWorkbenchService (apiCluster: ECSCluster, config: Config) {
+
+        const mcpWorkbenchImage = config.mcpWorkbenchConfig || {
+            baseImage: config.baseImage,
+            path: MCP_WORKBENCH_PATH,
+            type: EcsSourceType.ASSET
+        };
+
+        const mcpWorkbenchTaskDefinition = {
+            environment: {
+                RCLONE_CONFIG_S3_REGION: config.region,
+                MCPWORKBENCH_BUCKET: [config.deploymentName, config.deploymentStage, 'MCPWorkbench', config.accountNumber].join('-').toLowerCase(),
+            },
+            containerConfig: {
+                image: mcpWorkbenchImage,
+                healthCheckConfig: {
+                    command: ['CMD-SHELL', 'exit 0'],
+                    interval: 10,
+                    startPeriod: 30,
+                    timeout: 5,
+                    retries: 3
+                },
+                environment: {},
+                sharedMemorySize: 0,
+                privileged: true
+            },
+            containerMemoryReservationMiB: 1024,
+            applicationTarget: {
+                port: 8000,
+                priority: 80,
+                conditions: [{
+                    type: 'pathPatterns' as const,
+                    values: ['/v2/mcp/*']
+                }]
+            }
+        };
+
+        const { service } = apiCluster.addTask(ECSTasks.MCPWORKBENCH, mcpWorkbenchTaskDefinition);
+
+        this.createS3EventHandler(config, service);
+    }
+
+    private createS3EventHandler (config: any, workbenchService: Ec2Service) {
+        const s3EventHandlerRole = new iam.Role(this, 'S3EventHandlerRole', {
+            assumedBy: new iam.ServicePrincipal('lambda.amazonaws.com'),
+            inlinePolicies: {
+                'S3EventHandlerPolicy': new iam.PolicyDocument({
+                    statements: [
+                        new iam.PolicyStatement({
+                            effect: iam.Effect.ALLOW,
+                            actions: [
+                                'logs:CreateLogGroup',
+                                'logs:CreateLogStream',
+                                'logs:PutLogEvents'
+                            ],
+                            resources: [`arn:${config.partition}:logs:*:*:*`]
+                        }),
+                        new iam.PolicyStatement({
+                            effect: iam.Effect.ALLOW,
+                            actions: [
+                                'ecs:UpdateService',
+                                'ecs:DescribeServices',
+                                'ecs:DescribeClusters'
+                            ],
+                            resources: [
+                                `arn:${config.partition}:ecs:${config.region}:*:cluster/${workbenchService.cluster.clusterName}*`,
+                                `arn:${config.partition}:ecs:${config.region}:*:service/${workbenchService.cluster.clusterName}*/${workbenchService.serviceName}*`
+                            ]
+                        }),
+                        new iam.PolicyStatement({
+                            effect: iam.Effect.ALLOW,
+                            actions: [
+                                'ssm:GetParameter'
+                            ],
+                            resources: [
+                                `arn:${config.partition}:ssm:${config.region}:*:parameter${config.deploymentPrefix}/deploymentName`
+                            ]
+                        })
+                    ]
+                })
+            }
+        });
+
+        const s3EventHandlerLambda = new lambda.Function(this, 'S3EventHandlerLambda', {
+            runtime: getDefaultRuntime(),
+            handler: 'mcp_workbench.s3_event_handler.handler',
+            code: lambda.Code.fromAsset(config.lambdaPath ?? LAMBDA_PATH),
+            timeout: Duration.minutes(2),
+            role: s3EventHandlerRole,
+            environment: {
+                DEPLOYMENT_PREFIX: config.deploymentPrefix!,
+                API_NAME: 'MCPWorkbench',
+                ECS_CLUSTER_NAME: workbenchService.cluster.clusterName,
+                MCPWORKBENCH_SERVICE_NAME: workbenchService.serviceName
+            }
+        });
+
+        const rescanMcpWorkbenchRule = new events.Rule(this, 'RescanMCPWorkbenchRule', {
+            eventPattern: {
+                source: ['aws.s3', 'debug'],
+                detailType: [
+                    'Object Created',
+                    'Object Deleted'
+                ],
+                detail: {
+                    bucket: {
+                        name: [[config.deploymentName, config.deploymentStage, 'MCPWorkbench', config.accountNumber].join('-').toLowerCase()]
+                    }
+                }
+            },
+        });
+
+        rescanMcpWorkbenchRule.addTarget(new targets.LambdaFunction(s3EventHandlerLambda, {
+            retryAttempts: 2,
+            maxEventAge: Duration.minutes(5)
+        }));
     }
 }
