+++ test/cdk/mocks/MockApp.ts
@@ -166,8 +166,8 @@ export default class MockApp {
             vpc: networkingStack.vpc,
             restApiId: apiBaseStack.restApiId,
             rootResourceId: apiBaseStack.rootResourceId,
-            authorizerId: apiBaseStack.authorizer?.authorizerId || '',
-            apiCluster: serveStack.restApi.apiCluster
+            apiCluster: serveStack.restApi.apiCluster,
+            authorizer: apiBaseStack.authorizer
         });
 
         const stacks: cdk.Stack[] = [
